Reorganized project, setup.py
