+++ main.py
@@ -1,0 +1,6 @@
+import fire
+import raven.commands as commands
+
+# Initialize CLI
+
+fire.Fire(commands)
